+++ mrv2/lib/mrvWidgets/mrvMainWindow.cpp
@@ -109,6 +109,7 @@ namespace mrv
 #if defined(_WIN32)
         if (fl_win32_display())
         {
+            std::cerr << "load" << std::endl;
             Fl_Pixmap* pic = new Fl_Pixmap(viewer16_xpm);
             Fl_RGB_Image* rgb = new Fl_RGB_Image(pic);
             delete pic;